requirements.txt
selenium
pytest
python-docx
webdriver-manager